Updated grape-roar to 0.4.0
Fixes compatibility problems with representable 2.4+. An API change
meant that the current way of passing in the rack environment was
getting clobbered, and we were unable to access the context of the
request.

See
https://github.com/ruby-grape/grape-roar/pull/21/files
for the fix

and conversation in
https://github.com/slack-ruby/slack-ruby-bot-server/pull/43

Change properties to `public` methods

See #45 - https://github.com/slack-ruby/slack-ruby-bot-server/pull/45

https://github.com/trailblazer/representable/commit/74155a963b2db588515d2bcab53ab4dde41d21be
changed to use `public_send` for these types of methods.

Use danger format